hal(dma): widen DmaHal::Start ch from uint8_t to uint16_t
- hal(crypto): add timeout_ms parameter to ShaHal::Digest

+++ hal/include/backbone/dma_hal.h
@@ -45,7 +45,7 @@ class DmaHal {
      * @param ch Channel index.
      * @return HalStatus::Ok on success.
      */
-    HalStatus Start(uint8_t ch);
+    HalStatus Start(uint16_t ch);
 
     /**
      * @brief Stop a channel at the next beat boundary.
